Split backoffice review helper facets
